add ox with picture page and change class names and update main page button
