+++ app/src/main/java/com/example/testfolder/GamePictureActivity.kt
@@ -1,0 +1,109 @@
+package com.example.testfolder
+
+import android.media.MediaPlayer
+import android.os.Bundle
+import android.os.Handler
+import android.os.Looper
+import android.view.View
+import android.view.animation.AnimationUtils
+import android.widget.FrameLayout
+import android.widget.ImageView
+import android.widget.ProgressBar
+import android.widget.TextView
+import android.widget.Toast
+import androidx.activity.enableEdgeToEdge
+import androidx.appcompat.app.AppCompatActivity
+import androidx.core.view.ViewCompat
+import androidx.core.view.WindowInsetsCompat
+import com.google.firebase.auth.FirebaseUser
+
+class GamePictureActivity : BaseActivity() {
+    private lateinit var progressBar: ProgressBar
+    private var progressStatus = 0
+    private val handler = Handler(Looper.getMainLooper())
+    private lateinit var coinText: TextView
+    private lateinit var currentUser: FirebaseUser
+
+
+    private lateinit var roundImageView: ImageView
+    private lateinit var numberImageView: ImageView
+    private lateinit var finishedTextView: TextView
+
+    private lateinit var loadingLayout: View
+    private lateinit var loadingImage: ImageView
+    private lateinit var loadingText: TextView
+    private lateinit var qeustionImg: ImageView
+
+    private lateinit var obutton: FrameLayout
+    private lateinit var xbutton: FrameLayout
+
+    private lateinit var getCoin: MediaPlayer   //효과음 재생용 변수
+    private lateinit var wrong: MediaPlayer   //효과음 재생용 변수
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_game_picture)
+        progressBar = findViewById(R.id.progressBar1)
+
+        obutton = findViewById(R.id.o_btn)  // O버튼
+        xbutton = findViewById(R.id.x_btn) // X버튼
+        coinText = findViewById(R.id.coin_text)
+
+        roundImageView = findViewById(R.id.roundImageView)
+        numberImageView = findViewById(R.id.numberImageView)
+        finishedTextView = findViewById(R.id.finishedTextView)
+
+        loadingLayout = findViewById(R.id.loading_layout)
+        loadingImage = findViewById(R.id.loading_image)
+        loadingText = findViewById(R.id.loading_text)
+
+        qeustionImg = findViewById(R.id.qeustionImg)
+
+        getCoin = MediaPlayer.create(this,R.raw.coin)
+        wrong = MediaPlayer.create(this, R.raw.wrong)
+        // 로딩 이미지 회전 애니메이션 적용
+        val rotateAnimation = AnimationUtils.loadAnimation(this, R.anim.rotate)
+        loadingImage.startAnimation(rotateAnimation)
+        startLoadingTextAnimation()
+
+        qeustionImg.clipToOutline = true
+
+        // 답변 버튼 비활성화
+//        obutton.isEnabled = false
+//        xbutton.isEnabled = false
+
+
+        try {
+            currentUser = SingletonKotlin.getCurrentUser() ?: throw IllegalStateException("User authentication required.")
+        } catch (e: IllegalStateException) {
+            Toast.makeText(this, "SingletonKotlin is not initialized.", Toast.LENGTH_SHORT).show()
+            finish()
+            return
+        }
+
+        // 사용자 코인 불러오기
+        try {
+            SingletonKotlin.loadUserCoins(coinText)
+        } catch (e: IllegalStateException) {
+            Toast.makeText(this, "SingletonKotlin is not initialized.", Toast.LENGTH_SHORT).show()
+            finish()
+        }
+
+    }
+
+    private fun startLoadingTextAnimation() { //...을 로딩과 함께 애니메이션으로 움직이도록
+        var dotCount = 0
+        handler.post(object : Runnable {
+            override fun run() {
+                dotCount++
+                if (dotCount > 3) {
+                    dotCount = 0
+                }
+                val dots = ".".repeat(dotCount)
+                loadingText.text = "Pulling out the diary$dots"
+                handler.postDelayed(this, 500) // 500ms마다 업데이트
+            }
+        })
+    }
+
+}
