+++ app/src/main/java/com/example/testfolder/GamelistActivity.kt
@@ -1,13 +1,12 @@
 package com.example.testfolder
 
-import android.annotation.SuppressLint
 import android.content.Intent
 import android.os.Bundle
 import android.widget.Button
 import android.widget.LinearLayout
 import androidx.appcompat.app.AppCompatActivity
 
-class gamelistActivity : AppCompatActivity() {
+class GamelistActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_gamelist)
@@ -18,17 +17,23 @@ class gamelistActivity : AppCompatActivity() {
         val btn4 = findViewById<LinearLayout>(R.id.btn4)
         val btn5 = findViewById<LinearLayout>(R.id.btn5)
         val btn6 = findViewById<LinearLayout>(R.id.btn6)
+        val btn7 = findViewById<LinearLayout>(R.id.btn7)
+        val gRbtn = findViewById<Button>(R.id.game_record_btn)
 
         btn1.setOnClickListener {
-            val intent = Intent(this, gameLowActivity::class.java)
+            val intent = Intent(this, GameLowActivity::class.java)
             startActivity(intent)
         }
         btn2.setOnClickListener {
-            val intent = Intent(this, gameMidActivity::class.java)
+            val intent = Intent(this, GameMidActivity::class.java)
             startActivity(intent)
         }
         btn3.setOnClickListener {
-            val intent = Intent(this, gameHighActivity::class.java)
+            val intent = Intent(this, GameHighActivity::class.java)
+            startActivity(intent)
+        }
+        btn7.setOnClickListener {
+            val intent = Intent(this, GamePictureActivity::class.java)
             startActivity(intent)
         }
         btn4.setOnClickListener {
@@ -41,6 +46,10 @@ class gamelistActivity : AppCompatActivity() {
         }
         btn6.setOnClickListener {
             val intent = Intent(this, MinigameDescriptionSamepictureActivity::class.java)
+            startActivity(intent)
+        }
+        gRbtn.setOnClickListener {
+            val intent = Intent(this, OXGameRecordActivity::class.java)
             startActivity(intent)
         }
