+++ app/src/main/res/drawable/img_rounded.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android"
+    android:shape="rectangle">
+    <stroke
+        android:width="2dp"
+        android:color="@color/white"/>
+    <corners android:radius="5dp" />
+</shape>
