+++ app/src/main/res/layout/activity_game_high.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".gameHighActivity">
+    tools:context=".GameHighActivity">
 
     <LinearLayout
         android:id="@+id/coin_layout"
