+++ app/src/main/res/layout/activity_game_mid.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".gameMidActivity">
+    tools:context=".GameMidActivity">
 
     <LinearLayout
         android:id="@+id/coin_layout"
